Development into Main

See merge request ciis-capstone-project/winter-2022-2023/team-11/scida!56

+++ pages/api/deleteAccount.js
@@ -1,35 +1,55 @@
 //import the module to transform the posted data
 const querystring = require("querystring");
 //import mysql module to connect database
-const mysql = require("mysql");
+const mysql = require("mysql2");
+//Import nextauth to secure the api
+import { getSession } from "next-auth/react";
 
-export default function handler(req, res) {
-    // Get data submitted in request's body.
-    const body = req.body;
-  
-    const id = body.id;
+export default async (req, res) => {
+  const session = await getSession({ req });
 
-    //database information
-    const connection = mysql.createConnection({
-      host: "127.0.0.1",
-      user: "root",
-      password: "@UniKoeln123",
-      port: 3306,
-      database: "test_db",
-    });
-    //connect database
-    connection.connect();
-    //content query
-    connection.query(
-      "DELETE FROM accounts WHERE account_id=?",
-      [id],
-      (err, results, fields) => {
-        //error
-        res.status(200).json(`${results}`);
-        if (err) throw err;
-        res.end();
+  //Check if a session exists
+  if (session) {
+    //Try to recieve correct user role
+    var role;
+    try {
+      //Try ldap, if not existent do catch with local accounts
+      role = session.user.attributes.UniColognePersonStatus;
+    } catch {
+      role = session.user.account_role;
+    }
 
-        /* //data returned by database
+    //Check if users role is allowed to contact api, here role A (Admin i.e. Dekanat) and B (Beschäftigte i.e Sekretariat) is allowed
+    if (role === "scidaDekanat" || role === "scidaSekretariat") {
+      // Get data submitted in request's body.
+      const body = req.body;
+      const id = body.id;
+
+      //database information
+      const connection = mysql.createConnection({
+        host: "127.0.0.1",
+        user: "root",
+        password: "@UniKoeln123",
+        port: 3306,
+        database: "test_db",
+        timezone: "+00:00", //Use same timezone as in mysql database
+      });
+      //connect database
+      connection.connect();
+      //content query
+      connection.query(
+        "DELETE FROM accounts WHERE account_id=?",
+        [id],
+        (err, results, fields) => {
+          //error
+          res.status(200).json(`SUCCESS`);
+          if (err) throw err;
+          if (err) {
+            res.status(200).json(`FAIL CODE 4`);
+          }
+          res.end();
+
+          /* //data returned by database
 			//if no such data existed in this database,
 			//a empty array with be returned(looks like this: []),
 			//and length of results would be zero 
@@ -40,9 +60,9 @@ export default function handler(req, res) {
 				res.write('you logged in!');
 				res.end();
 			} */
-    }
-    );
-    /*
+        }
+      );
+      /*
       connection.query(
         "select * from account",
         (err, results, fields) => {
@@ -66,6 +86,16 @@ export default function handler(req, res) {
         
     );*/
 
-    // disconnect database
-    connection.end();
+      // disconnect database
+      connection.end();
+    }
+    //Return unAUTHORIZED if wrong role
+    else {
+      res.status(401).json({ error: "Unauthorized user -> Wrong role" });
+    }
   }
+  //Return unAUTHENTICATED if not logged in
+  else {
+    res.status(401).json({ error: "Unauthenticated user -> Not logged in" });
+  }
+};
